+++ codex-rs/core-api/Cargo.toml
@@ -25,6 +25,5 @@ codex-features = { workspace = true }
 codex-login = { workspace = true }
 codex-model-provider-info = { workspace = true }
 codex-models-manager = { workspace = true }
-codex-plugins-extension = { workspace = true }
 codex-protocol = { workspace = true }
 codex-utils-absolute-path = { workspace = true }
