+++ codex-rs/core/src/prompt_debug.rs
@@ -20,8 +20,7 @@ use crate::session::turn::built_tools;
 use crate::state_db_bridge::StateDbHandle;
 use crate::thread_manager::ThreadManager;
 use crate::thread_manager::thread_store_from_config;
-use codex_extension_api::ExtensionRegistry;
-use codex_extension_api::ExtensionRegistryBuilder;
+use codex_extension_api::empty_extension_registry;
 
 /// Build the model-visible `input` list for a single debug turn.
 #[doc(hidden)]
@@ -51,7 +50,7 @@ pub async fn build_prompt_input(
                 .await
                 .map_err(|err| CodexErr::Fatal(err.to_string()))?,
         ),
-        plugins_extension_registry(),
+        empty_extension_registry(),
         /*analytics_events_client*/ None,
         thread_store,
         state_db.clone(),
@@ -105,10 +104,4 @@ pub(crate) async fn build_prompt_input_from_session(
     );
 
     Ok(prompt.get_formatted_input())
-}
-
-fn plugins_extension_registry() -> Arc<ExtensionRegistry<Config>> {
-    let mut builder = ExtensionRegistryBuilder::new();
-    codex_plugins_extension::install(&mut builder);
-    Arc::new(builder.build())
 }
