+++ codex-rs/core/src/tools/mod.rs
@@ -4,7 +4,6 @@ pub(crate) mod events;
 pub(crate) mod handlers;
 pub(crate) mod hook_names;
 pub(crate) mod hosted_spec;
-pub(crate) mod installable_plugins;
 pub(crate) mod network_approval;
 pub(crate) mod orchestrator;
 pub(crate) mod parallel;
