+++ codex-rs/thread-manager-sample/src/main.rs
@@ -53,10 +53,10 @@ use codex_core_api::UserInput;
 use codex_core_api::WebSearchMode;
 use codex_core_api::arg0_dispatch_or_else;
 use codex_core_api::built_in_model_providers;
+use codex_core_api::empty_extension_registry;
 use codex_core_api::find_codex_home;
 use codex_core_api::init_state_db;
 use codex_core_api::item_event_to_server_notification;
-use codex_core_api::plugins_extension_registry;
 use codex_core_api::resolve_installation_id;
 use codex_core_api::set_default_originator;
 use codex_core_api::thread_store_from_config;
@@ -123,7 +123,7 @@ async fn run_main(arg0_paths: Arg0DispatchPaths) -> anyhow::Result<()> {
         auth_manager,
         SessionSource::Exec,
         environment_manager,
-        plugins_extension_registry(),
+        empty_extension_registry(),
         /*analytics_events_client*/ None,
         Arc::clone(&thread_store),
         state_db,
